Added more unit tests.
Refactored the GET and DELETE APIs to not require Content-Type header

+++ src/main/java/io/company/brewcraft/repository/ExtendedRepository.java
@@ -3,7 +3,7 @@ package io.company.brewcraft.repository;
 public interface ExtendedRepository<ID> {
     boolean existsByIds(Iterable<ID> ids);
 
-    long deleteByIds(Iterable<ID> ids);
+    int deleteByIds(Iterable<ID> ids);
 
-    long deleteOneById(ID id);
+    int deleteOneById(ID id);
 }
